refactor: renamed stac.py to download.py for more explicit. fixed tests
